Update project group ID
