+++ src/main/java/store/vxdesign/apps/cryptography/algorithms/mh/MerkleHellmanResultType.java
@@ -1,0 +1,17 @@
+/*
+ * Copyright 2017 Roman Mashenkin
+ * Licensed under the Apache License, Version 2.0
+ */
+package store.vxdesign.apps.cryptography.algorithms.mh;
+
+/**
+ * Merkle-Hellman knapsack cryptosystem enumeration class of results.
+ *
+ * @author Roman Mashenkin
+ * @since 20.10.2017
+ */
+public enum MerkleHellmanResultType {
+
+    TEXT_INPUT, BINARY_INPUT, HEX_INPUT, PRIVATE_KEY, PUBLIC_KEY, OUTPUT_SUMS, INPUT_SUMS, TEXT_OUTPUT, BINARY_OUTPUT,
+    HEX_OUTPUT
+}
