--- src/main/java/store/vxdesign/cryptography/algorithms/des/DataEncryptionStandardResultType.java
@@ -1,16 +1,0 @@
-/*
- * Copyright 2017 Roman Mashenkin
- * Licensed under the Apache License, Version 2.0
- */
-package store.vxdesign.cryptography.algorithms.des;
-
-/**
- * Data Encryption Standard enumeration class of results.
- *
- * @author Roman Mashenkin
- * @since 18.10.2017
- */
-enum DataEncryptionStandardResultType {
-
-    TEXT_INPUT, BINARY_INPUT, HEX_INPUT, TEXT_KEY, BINARY_KEY, HEX_KEY, TEXT_OUTPUT, BINARY_OUTPUT, HEX_OUTPUT
-}
