--- src/main/java/store/vxdesign/cryptography/framework/files/FileExtension.java
@@ -1,16 +1,0 @@
-/*
- * Copyright 2017 Roman Mashenkin
- * Licensed under the Apache License, Version 2.0
- */
-package store.vxdesign.cryptography.framework.files;
-
-/**
- * Class of file extension types.
- *
- * @author Roman Mashenkin
- * @since 19.10.2017
- */
-public enum FileExtension {
-
-    TEXT, KEY
-}
